pyproject.toml
[tool.poetry]
name = "mcarchive"
version = "0.1.0"
description = "backend for MCArchive"
authors = ["tropicalalgae Your email"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/tropical-algae/MCArchive-backend.git"


[[tool.poetry.source]]
name = "proxy-source"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.0"
uvicorn = "0.30.5"
pydantic = { extras = ["email"], version = "^2.9.2" }
requests = "^2.32.0"
loguru = "^0.7.0"
joblib = "^1.2.0"
poethepoet = "^0.27.0"
pydantic-settings = "^2.4.0"
colorlog = "^6.8.2"
pytz = "^2024.1"
openai = "^1.38.0"
pyyaml = "^6.0.1"
pymysql = "^1.1.1"
sqlalchemy = "^2.0.32"
passlib = "^1.7.4"
python-jose = "^3.3.0"
python-multipart = "^0.0.9"
sqlmodel = "^0.0.22"
cachetools = "^6.2.1"
alibabacloud-oss-v2 = "^1.2.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2"
pytest-cov = "^5.0.0"
pytest-ordering = "^0.6"
black = { version = "^24.3", allow-prereleases = true }
autopep8 = "^2.0.0"
ipdb = "^0.13.0"
mypy = "^1.11.1"
ruff = "^0.5.6"
types-pyyaml = "^6.0.12.20240724"
types-pytz = "^2024.1.0.20240417"
sqlacodegen = "3.0.0rc5"
pre-commit = "^4.3.0"
types-cachetools = "^6.2.0.20251022"


[tool.poetry.group.release.dependencies]
python-semantic-release = "^10.3.1"


[tool.poetry.group.aws]
optional = true


[tool.poetry.group.aws.dependencies]
mangum = "^0.17.0"


[[tool.poetry.packages]]
include = "mcarchive"
from = "src"


[tool.poe]
include = "poe_tasks.toml"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"


[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
  \.toml
  |\.sh
  |\.git
  |\.ini
  |Dockerfile
  |Jenkinfile
)/
'''
